requirements.txt
influxdb-client>=1.36.0
python-dotenv>=1.0.0
click>=8.0.0
tqdm>=4.64.0